refactor: unify submission pending queries
